Fix two sanitizer findings: a one-byte read before g_tmp on short macro labels, and a signed shift overflow in wlalink's READ_T
decode.c: the "\@" check on a label inside a macro reads g_tmp[g_ss - 3]
without checking that the label is at least two characters long. Any
one-character label inside a macro ("-:", "+:", "x:") reads one byte
before the calloc'd buffer (AddressSanitizer heap-buffer-overflow,
found on OpenSNES' snesmod.asm and object.asm). Guard on g_ss >= 3.

wlalink/{analyze,check,listfile}.c: READ_T computes t[0] << 24 on an
unsigned char promoted to int, which overflows when the high byte is
128 or more (UndefinedBehaviorSanitizer "left shift of 144 by 24 places
cannot be represented in type 'int'", on a section checksum). Widen the
top byte to unsigned int before shifting; the sum is unchanged on every
platform wla-dx supports.

+++ wlalink/analyze.c
@@ -19,7 +19,7 @@
 
 
 /* read an integer from t */
-#define READ_T (t[3] + (t[2] << 8) + (t[1] << 16) + (t[0] << 24)); t += 4;
+#define READ_T (t[3] + (t[2] << 8) + (t[1] << 16) + ((unsigned int)t[0] << 24)); t += 4;
 
 /* read a double from t */
 #define READ_DOU {                              \
@@ -2328,7 +2328,7 @@ int obtain_rombanks(void) {
     if (g_obj_tmp->format == WLA_VERSION_OBJ) {
 
       t = g_obj_tmp->data + OBJ_ROMBANKS;
-      k = t[3] + (t[2] << 8) + (t[1] << 16) + (t[0] << 24);
+      k = t[3] + (t[2] << 8) + (t[1] << 16) + ((unsigned int)t[0] << 24);
 
       g_obj_tmp->rom_banks = k;
